[chore#zsh] Add check folders during installing oh-my-zsh

+++ zsh/setup_oh_my_zsh.sh
@@ -1,18 +1,46 @@
 #!/usr/bin/env bash
 
-echo "Installing oh-my-zsh"
+if [ ! -d "ZSH" ]; then
+    echo "Installing oh-my-zsh.."
+    sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+    echo "DONE!"
+fi
 
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 
-rm -v "$HOME/.zshrc"
+if [ -f "$HOME/.zshrc" ]; then
+    rm -v "$HOME/.zshrc"
+fi
 
 
-echo "Installing pluings.."
+if [ ! -d "$ZSH_CUSTOM/themes/powerlevel10k" ]; then
+    echo "Installing powerlevel10k"
+    git clone --depth=1 https://github.com/romkatv/powerlevel10k.git $ZSH_CUSTOM/themes/powerlevel10k
+    echo "DONE!"
+fi
 
-git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
+echo "Installing pluings..."
 
-git clone https://github.com/jeffreytse/zsh-vi-mode $ZSH_CUSTOM/plugins/zsh-vi-mode
-git clone https://github.com/zsh-users/zsh-autosuggestions.git $ZSH_CUSTOM/plugins/zsh-autosuggestions
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git $ZSH_CUSTOM/plugins/zsh-syntax-highlighting
-git clone https://github.com/zdharma-continuum/fast-syntax-highlighting.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/plugins/fast-syntax-highlighting
-git clone --depth 1 -- https://github.com/marlonrichert/zsh-autocomplete.git $ZSH_CUSTOM/plugins/zsh-autocomplete
+PLUGINS="$ZSH_CUSTOM/plugins"
+
+if [ ! -d "$PLUGINS/zsh-vi-mode" ]; then
+    echo "Install vi-mode"
+    git clone https://github.com/jeffreytse/zsh-vi-mode "$PLUGINS/zsh-vi-mode"
+fi
+if [ ! -d "$PLUGINS/zsh-autosuggestions" ]; then
+    echo "Install autosuggestions"
+    git clone https://github.com/zsh-users/zsh-autosuggestions.git "$PLUGINS/zsh-autosuggestions"
+fi
+if [ ! -d "$PLUGINS/zsh-syntax-highlighting" ]; then
+    echo "Install syntax-highlighting"
+    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git "$PLUGINS/zsh-syntax-highlighting"
+fi
+if [ ! -d "$PLUGINS/fast-syntax-highlighting" ]; then
+    echo "Install fast-syntax-highlighting"
+    git clone https://github.com/zdharma-continuum/fast-syntax-highlighting.git "$PLUGINS/fast-syntax-highlighting"
+fi
+if [ ! -d "$PLUGINS/zsh-autocomplete" ]; then
+    echo "Install autocomplete"
+    git clone --depth 1 -- https://github.com/marlonrichert/zsh-autocomplete.git "$PLUGINS/zsh-autocomplete"
+fi
+
+echo "DONE"
